Se implementa test del primer step del escenario 'Intentar eliminar un grupo que no existe'

+++ src/main/java/ar/com/grupoesfera/repartir/services/GruposService.java
@@ -70,6 +70,12 @@ public class GruposService {
     }
 
     public void eliminar(Long id) {
+        Optional<Grupo> grupoBuscado = repository.findById(id);
+
+        if (!grupoBuscado.isPresent()) {
+            throw new GrupoNoEncontradoException();
+        }
+
         repository.deleteById(id);
     }
